QuickSearch

Documentation

buildQuickSearch

Arguments

:: (Hashable uid, Eq uid)

=> [(Text, uid)]
List of entries to be searched

-> QuickSearch uid
QuickSearch object holding token partitions

Given a list of pairs of (T.Text, uid) to be searched, create a QuickSearch object.

matchesWithThreshold

Arguments

:: (Hashable uid, Eq uid)

=> QuickSearch uid
QuickSearch object holding token partitions

-> Int
Threshold score above which to return results

-> Scorer
String similarity function of type (Text -> Text -> Ratio Int)

-> Text
String to be searched

-> [Match (Entry uid)]
Top N most similar entries

Given a QuickSearch object, scorer, and string, return all matches with a score greater than the given threshold.

topNMatches

Arguments

:: (Hashable uid, Eq uid)

=> QuickSearch uid
QuickSearch object holding token partitions

-> Int
N: Number of results to return

-> Scorer
String similarity function of type (Text -> Text -> Ratio Int)

-> Text
String to be searched

-> [Match (Entry uid)]
Top N most similar entries

Given a QuickSearch object, scorer, and string, return the top N matches.

batch

Arguments

:: (Hashable uid1, Eq uid1, Hashable uid2, Eq uid2)

=> (QuickSearch uid2 -> Int -> Scorer -> Text -> [Match (Entry uid2)])
A match retrieval function, such as topNMatches

-> QuickSearch uid2
QuickSearch object holding token partitions

-> Int
The reference number for the match retrieval function. N for topNMatches, threshold for matchesWithThreshold

-> Scorer
String similarity function of type (Text -> Text -> Ratio Int)

-> [(Text, uid1)]
List of entries to be processed

-> [(Entry uid1, [Match (Entry uid2)])]
List of entries and the results returned for each.

Turn a match retrieval function into one that works on lists of entries.

batchTopNMatches

Arguments

:: (Hashable uid1, Eq uid1, Hashable uid2, Eq uid2)

=> QuickSearch uid2
QuickSearch object holding token partitions

-> Int
N: Number of results to return

-> Scorer
String similarity function of type (Text -> Text -> Ratio Int)

-> [(Text, uid1)]
List of entries to be processed

-> [(Entry uid1, [Match (Entry uid2)])]
List of entries and up to the top N matches for each.

Version of topNMatches that processes lists of entries instead of strings.

batchMatchesWithThreshold

Arguments

:: (Hashable uid1, Eq uid1, Hashable uid2, Eq uid2)

=> QuickSearch uid2
QuickSearch object holding token partitions

-> Int
N: Number of results to return

-> Scorer
String similarity function of type (Text -> Text -> Ratio Int)

-> [(Text, uid1)]
List of entries to be processed

-> [(Entry uid1, [Match (Entry uid2)])]
List of entries and their matches above the score threshold.

Version of matchesWithThreshold that processes lists of entries instead of strings.

type Token = Text

type Score = Int

type Scorer = Text -> Text -> Ratio Int

data Entry uid

Constructors

Entry

Fields

entryName :: Text

entryUID :: uid

data Match a

Constructors

Match

Fields

matchScore :: Score

matchEntry :: a

data QuickSearch uid

Constructors

QuickSearch [Entry uid] (HashMap Token (HashSet uid))