QuickSearch.Internal.Filter

Documentation

buildTokenPartitions

Arguments

:: (Hashable uid, Eq uid)

=> [Entry uid]
List of entries

-> HashMap Token (HashSet uid)
A map of Token -> [uids]

Given the list of entries to be held by QuickSearch, return a HashMap keyed on tokens from the strings in the entries, where the associated HashMap value is the list of uids of entries containing the token.

getSearchPartition

Arguments

:: (Hashable uid, Eq uid)

=> Text
Target string

-> HashMap Token (HashSet uid)
HashMap associating tokens with sets of uids

-> HashSet uid
The union of sets of associated uids.

Given a target string and a Token HashMap, return the union of sets of uids associated with the tokens in the target string

type Token = Text

data Entry uid

Constructors

Entry

Fields

entryName :: Text

entryUID :: uid

first :: (Text -> Text) -> Entry uid -> Entry uid