QuickSearch.Internal.Matcher

Documentation

type Score = Int

type Scorer = Text -> Text -> Ratio Int

newtype Match score entry

Structure associating a Score with an Entry, for holding search results

Constructors

Match (score, entry)

newtype QuickSearch uid

List of entries to be searched and a HashMap associating tokens with HashSets of UIDs related to entries containing the tokens.

Constructors

QuickSearch ([Entry Text uid], HashMap Token (HashSet uid))

scoreMatches

Arguments

:: (Hashable uid, Eq uid)

=> Text
Name or other string to be searched

-> QuickSearch uid
The QuickSearch object to be used

-> Scorer
A string distance function of type (Text -> Text -> Ratio Int)

-> [Match Score (Entry Text uid)]
A list of possible matches

Ignore the linter here, this is a performance thing

Given a string to search, a QuickSearch object, and a similarity function, returns potential matches contained in the QuickSearch filters and their associated scores, in descending order by score.

matchScore :: Match Score (Entry name uid) -> Score

Score accessor for Match

matchEntry :: Match Score (Entry name uid) -> Entry name uid

Entry accessor for Match

quickSearchEntries :: (Hashable uid, Eq uid) => QuickSearch uid -> [Entry Text uid]

Entry name uid

accessor for QuickSearch

quickSearchTokenFilter :: (Hashable uid, Eq uid) => QuickSearch uid -> HashMap Token (HashSet uid)

tokenFilter accessor for QuickSearch