QuickSearch.Matcher

Documentation

scoreMatches

Arguments

:: (Hashable uid, Eq uid)

=> Text
Name or other string to be searched

-> QuickSearch uid
The QuickSearch object to be used

-> Scorer
A string distance function of type (Text -> Text -> Ratio Int)

-> [Match (Entry uid)]
A list of possible matches and their scores

Ignore the linter here, this is a performance thing

Given a string to search, a QuickSearch object, and a similarity function, returns potential matches contained in the QuickSearch filters and their associated scores, in descending order by score.

type Token = Text

type Score = Int

type Scorer = Text -> Text -> Ratio Int

data Match a

Constructors

Match

Fields

matchScore :: Score

matchEntry :: a

data QuickSearch uid

Constructors

QuickSearch [Entry uid] (HashMap Token (HashSet uid))