QuickSearch.String.OneShot

Documentation

oneShot

Arguments

:: (Hashable uid1, Eq uid1, Hashable uid2, Eq uid2)

=> (QuickSearch uid2 -> Int -> Scorer -> String -> [Match (SEntry uid2)])
Match retrieval function to be converted into a one-shot

-> Int
The reference number for the match retrieval function.

-> [(String, uid1)]
List of entries to be processed

-> [(String, uid2)]
List of entries making up the search space

-> Scorer
Similarity function with type (Text -> Text -> Ratio Int)

-> [(SEntry uid1, [Match (SEntry uid2)])]
List of entries and their matches.

Turn a match retrieval function into a one-shot batch function. Instead of creating a QuickSearch for reuse, this creates it in the background and discards it when done.

oneShotTopNMatches

Arguments

:: (Hashable uid1, Eq uid1, Hashable uid2, Eq uid2)

=> Int
N: Number of matches to return

-> [(String, uid1)]
List of entries to be processed

-> [(String, uid2)]
List of entries making up the search space

-> Scorer
Similarity function with type (Text -> Text -> Ratio Int)

-> [(SEntry uid1, [Match (SEntry uid2)])]
List of entries and up to N of the best matches.

One-shot version of topNMatches. Builds the QuickSearch in the background and discards it when finished.

oneShotMatchesWithThreshold

Arguments

:: (Hashable uid1, Eq uid1, Hashable uid2, Eq uid2)

=> Int
Score threshold above which to return matches

-> [(String, uid1)]
List of entries to be processed

-> [(String, uid2)]
List of entries making up the search space

-> Scorer
Similarity function with type (Text -> Text -> Ratio Int)

-> [(SEntry uid1, [Match (SEntry uid2)])]
List of entries and their matches above the score threshold.

One-shot version of matchesWithThreshold. Builds the QuickSearch in the background and discards it when finished.